test(e2e): split TOC markup assertions into L3 dist reads
smoke-toc.spec.ts keeps only what needs a real browser: viewport
visibility and scroll-spy behavior. The static markup checks
(h2/h3 heading link presence, h3 indent class) move to the new
smoke-toc-markup.spec.ts as dist reads — same assertions, scoped to
the extracted <nav aria-label="Table of contents"> element to match
the original locator scope.

Part of zudolab/zudo-doc#2537.
